+++ pkg/service/project.go
@@ -22,6 +22,7 @@ import (
 	"database/sql"
 	"fmt"
 
+	"github.com/apache/airavata-custos/pkg/events"
 	"github.com/apache/airavata-custos/pkg/models"
 )
 
@@ -64,6 +65,8 @@ func (s *Service) CreateProject(ctx context.Context, project *models.Project) (*
 	}); err != nil {
 		return nil, fmt.Errorf("create project: %w", err)
 	}
+
+	s.eventBus.Publish(events.ProjectCreateEvent, project)
 	return project, nil
 }
 
@@ -110,6 +113,8 @@ func (s *Service) UpdateProject(ctx context.Context, project *models.Project) er
 	}); err != nil {
 		return fmt.Errorf("update project: %w", err)
 	}
+
+	s.eventBus.Publish(events.ProjectUpdateEvent, project)
 	return nil
 }
 
@@ -118,10 +123,19 @@ func (s *Service) DeleteProject(ctx context.Context, id string) error {
 	if id == "" {
 		return fmt.Errorf("%w: project id is required", ErrInvalidInput)
 	}
+	project, err := s.projs.FindByID(ctx, id)
+	if err != nil {
+		return fmt.Errorf("lookup project: %w", err)
+	}
+	if project == nil {
+		return ErrNotFound
+	}
 	if err := s.inTx(ctx, func(tx *sql.Tx) error {
 		return s.projs.Delete(ctx, tx, id)
 	}); err != nil {
 		return fmt.Errorf("delete project: %w", err)
 	}
+
+	s.eventBus.Publish(events.ProjectDeleteEvent, project)
 	return nil
 }
